FY-131/refactor: ProfileInfo 컴포넌트 분리
