+++ src/components/profile/ProfileInfoTextLine.js
@@ -1,0 +1,28 @@
+import React, { useEffect, useState } from "react";
+
+const ProfileInfoTextLine = ({ label, value, isEditing = false, onChange }) => {
+  const [inputValue, setInputValue] = useState(value);
+
+  useEffect(() => {
+    setInputValue(value);
+  }, [value, isEditing]);
+
+  const handleChange = (e) => {
+    const newValue = e.target.value;
+    setInputValue(newValue);
+    onChange(newValue);
+  };
+
+  return (
+    <div className="row">
+      <div className="col">
+        <label>{label}: </label>
+      </div>
+      <div className="col">
+        {isEditing ? <input type="text" value={inputValue} onChange={handleChange} /> : <span>{inputValue}</span>}
+      </div>
+    </div>
+  );
+};
+
+export default ProfileInfoTextLine;
